% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoReport}
\alias{autoReport}
\title{autoReport}
\usage{
autoReport(contig.folder, ref.folder, name.file, out.folder = "output",
  filter = FALSE, percentage = 1)
}
\arguments{
\item{contig.folder}{contig folder}

\item{ref.folder}{reference folder}

\item{name.file}{name file}

\item{out.folder}{output folder}

\item{filter}{logical}

\item{percentage}{filter percentage}
}
\description{
auto report sequence alignment and consensus sequence
}
\author{
ygc
}

