% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_summary_file}
\alias{convert_summary_file}
\title{convert_summary_file}
\usage{
convert_summary_file(summary_file, out_file = "summary2.csv")
}
\arguments{
\item{summary_file}{summary_file}

\item{out_file}{out_file}
}
\description{
convert 454 summary file for generateGapFile function
}
\author{
Guangchuang Yu
}

