% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAlign.R
\name{doAlign}
\alias{doAlign}
\title{doAlign}
\usage{
doAlign(x)
}
\arguments{
\item{x}{file names of 454 contig, reference and mira sequences}
}
\value{
fasta S3 object
}
\description{
align 454 contigs and mira sequence to reference sequence.
}
\author{
ygc
}

