% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAlign.R
\name{doAlign2}
\alias{doAlign2}
\title{doAlign2}
\usage{
doAlign2(files)
}
\arguments{
\item{files}{fasta files}
}
\value{
alignment
}
\description{
align multiple sequence stored in several fasta files.
}
\author{
ygc
}

