% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCBI.R
\name{download_genbank_acc}
\alias{download_genbank_acc}
\title{download_genbank_acc}
\usage{
download_genbank_acc(acc, db = "nuccore", outfile = NULL)
}
\arguments{
\item{acc}{accession number}

\item{db}{supported db, currently 'nuccore'}

\item{outfile}{output file, by default, acc_number.gb}
}
\description{
download genbank (.gb) file by accession number
}
\author{
Guangchuang Yu
}

