% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCBI.R
\name{download_genbank_gi}
\alias{download_genbank_gi}
\title{download_genbank_gi}
\usage{
download_genbank_gi(gi, db = "nuccore", outfile = NULL)
}
\arguments{
\item{gi}{gi number}

\item{db}{supported db, currently 'nuccore'}

\item{outfile}{output file, by default, gi_number.gb}
}
\description{
download genbank (.gb) file by GI number
}
\author{
Guangchuang Yu
}

