% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_netNglyc}
\alias{extract_netNglyc}
\title{extract_netNglyc}
\usage{
extract_netNglyc(infile, outfile, cutoff = 0.5, simplify = TRUE)
}
\arguments{
\item{infile}{output of NetNGlyc}

\item{outfile}{output file}

\item{cutoff}{cutoff}

\item{simplify}{boolean to remove null column}
}
\value{
data.frame
}
\description{
extract N-glyco sites from output of NetNGlyc 1.0
}
\author{
Guangchuang Yu
}

