% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fileMerge}
\alias{fileMerge}
\title{fileMerge}
\usage{
fileMerge(files = NULL, pattern = "txt$", output,
  removeColnames = "Frequency", removeRownames = "Frequency")
}
\arguments{
\item{files}{input files}

\item{pattern}{input file pattern}

\item{output}{output file}

\item{removeColnames}{rows to remove}

\item{removeRownames}{columns to remove}
}
\value{
data.frame
}
\description{
merge files and generate annotation file for treeAnno
}
\details{
currently support merging output of NetNGlyc server (http://www.cbs.dtu.dk/services/NetNGlyc/)
}
\author{
Guangchuang Yu
}

