% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateGapFile}
\alias{generateGapFile}
\title{generateGapFile}
\usage{
generateGapFile(out.folder = "output", ref.folder = "Ref", read.fileName,
  sample.size = 200, gap.file = "gaps.txt")
}
\arguments{
\item{out.folder}{output folder of autoReport, that contains consensus sequences}

\item{ref.folder}{reference folder}

\item{read.fileName}{list of file names of read file}

\item{sample.size}{sample size}

\item{gap.file}{output gap file name}
}
\description{
generating gap file
}
\author{
ygc
}

