% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawMap.R
\name{get.map}
\alias{get.map}
\title{get.map}
\usage{
get.map(map = "china")
}
\arguments{
\item{map}{only 'china' is supported}
}
\value{
data.frame
}
\description{
get map coordination
}
\author{
Guangchuang Yu
}

