% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NGS.R
\docType{methods}
\name{getPosition}
\alias{getPosition}
\alias{getPosition,NGS,ANY-method}
\title{getPosition method}
\usage{
getPosition(x, ...)

getPosition(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameters}
}
\value{
NGS positions
}
\description{
getPosition method

getPosition method
}
\author{
Guangchuang Yu
}

