% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NGS.R
\docType{methods}
\name{getSite}
\alias{getSite}
\alias{getSite,NGS,ANY-method}
\title{getSite method}
\usage{
getSite(x, ...)

getSite(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameters}
}
\value{
NGS site
}
\description{
getSite method

getSite method
}
\author{
Guangchuang Yu
}

