% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hamming.BStringSet}
\alias{hamming.BStringSet}
\title{hamming}
\usage{
\method{hamming}{BStringSet}(seq1, seq2, indel = FALSE, ...)
}
\arguments{
\item{seq1}{sequence 1}

\item{seq2}{sequence 2}

\item{indel}{include indel or not}

\item{...}{additional parameter}
}
\value{
numeric
}
\description{
hamming
}
\author{
ygc
}

