% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{hamming}
\alias{hamming}
\title{hamming}
\usage{
hamming(seq1, seq2, indel = FALSE, ...)
}
\arguments{
\item{seq1}{sequence 1}

\item{seq2}{sequence 2}

\item{indel}{include indel or not}

\item{...}{additional parameter}
}
\description{
hamming
}

