% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylip.R
\name{mergePhy}
\alias{mergePhy}
\title{mergePhy}
\usage{
mergePhy(p1, p2, outfile)
}
\arguments{
\item{p1}{phylip file 1}

\item{p2}{phylip file 2}

\item{outfile}{output file}
}
\description{
merge two phylip files into one.
}
\author{
ygc
}

