% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NGS.R
\docType{class}
\name{NGS-class}
\alias{NGS-class}
\alias{getPosition,NGS-method}
\alias{getSite,NGS-method}
\alias{setNGS_,NGS-method}
\alias{show,NGS-method}
\alias{toAA,NGS-method}
\alias{toCodon,NGS-method}
\alias{toPosition,NGS-method}
\alias{toString,NGS-method}
\title{Class "NGS"
This class stores information for N-glycosylation sites}
\description{
Class "NGS"
This class stores information for N-glycosylation sites
}
\section{Slots}{

\describe{
\item{\code{stringSet}}{BStringSet}

\item{\code{start}}{start position}

\item{\code{characterSet}}{sequence character vector}

\item{\code{codonSet}}{codon of sequences}

\item{\code{AASet}}{translate codonSet to AASet}

\item{\code{position}}{NGS position}
}}
\author{
GuangchuangYu \url{http://ygc.name}
}
\keyword{classes}

