% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAnno}
\alias{plotAnno}
\title{plotAnno}
\usage{
plotAnno(anno1, anno2 = NULL, xlab = "Position", ylab = "Potential",
  title = "")
}
\arguments{
\item{anno1}{annotation data.frame, require}

\item{anno2}{annotation data.frame, optional}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}
}
\value{
ggplot2 object
}
\description{
plot functional site annotation
}
\author{
Guangchuang
}

