% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawMap.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(mapdata, xlim = c(110, 125), ylim = c(20, 35), location, lineWidth,
  maxLineSize = 3, color, colorIntensity, arrowHeadSize = 1,
  matrixFlag = "upper", mapcol = "darkgrey", ggmap = FALSE,
  maptype = "hybrid")
}
\arguments{
\item{mapdata}{map coordination}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{location}{location matrix}

\item{lineWidth}{linewidth matrix}

\item{maxLineSize}{maxLineSize}

\item{color}{color matrix}

\item{colorIntensity}{color intensity matrix}

\item{arrowHeadSize}{arrow head size}

\item{matrixFlag}{one of upper or lower}

\item{mapcol}{color of background map}

\item{ggmap}{whether use ggmap or not}

\item{maptype}{map type of ggmap}
}
\value{
ggplot2 figure
}
\description{
plot map
}
\author{
Guangchuang Yu
}

