% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawMap.R
\name{plotMap2}
\alias{plotMap2}
\title{plotMap}
\usage{
plotMap2(mapdata, xlim = c(110, 125), ylim = c(20, 35), location_file,
  lineWidth_file, maxLineSize = 3, color_file, colorIntensity_file,
  arrowHeadSize = 1, matrixFlag = "upper", mapcol = "darkgrey",
  ggmap = FALSE, maptype = "hybrid")
}
\arguments{
\item{mapdata}{map coordination}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{location_file}{location file}

\item{lineWidth_file}{line width file}

\item{maxLineSize}{maxLineSize}

\item{color_file}{color file}

\item{colorIntensity_file}{color intensity file}

\item{arrowHeadSize}{arrow head size}

\item{matrixFlag}{one of upper or lower}

\item{mapcol}{color of background map}

\item{ggmap}{whether use ggmap or not}

\item{maptype}{map type of ggmap}

\item{location}{location matrix}

\item{lineWidth}{linewidth matrix}

\item{color}{color matrix}

\item{colorIntensity}{color intensity matrix}
}
\value{
ggplot2 figure
}
\description{
plot map
}
\author{
Guangchuang Yu
}

