% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeAnno.pml.R
\name{pmlToSeq}
\alias{pmlToSeq}
\title{pmlToSeq}
\usage{
pmlToSeq(pml, includeAncestor = TRUE, output = "seq.fa")
}
\arguments{
\item{pml}{pml object}

\item{includeAncestor}{logical}

\item{output}{out fasta file}
}
\value{
XStringSet
}
\description{
convert pml object to fasta file or XStringSet object
}
\author{
ygc
}

