% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{printAlignedSeq}
\alias{printAlignedSeq}
\title{printAlignedSeq}
\usage{
printAlignedSeq(aln, window = 80)
}
\arguments{
\item{aln}{alignment}

\item{window}{window}
}
\description{
print aligned sequences
}
\author{
ygc
}

