% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printConsensus}
\alias{printConsensus}
\title{printConsensus}
\usage{
printConsensus(aln, window = 10)
}
\arguments{
\item{aln}{alignment}

\item{window}{window of ambiguous base}
}
\value{
character
}
\description{
print the consensus sequence
}
\author{
ygc
}

