% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCBI.R
\name{read.genbank}
\alias{read.genbank}
\title{read.genbank}
\usage{
read.genbank(gbfile)
}
\arguments{
\item{gbfile}{genbank file}
}
\value{
data.frame
}
\description{
parse genbank file and extract infomation
(not a universal parse, just extract infomation of my own needs)
}
\author{
Guangchuang Yu
}

