% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phylip.R
\name{read.phylip}
\alias{read.phylip}
\title{read.phylip}
\usage{
read.phylip(file)
}
\arguments{
\item{file}{phylip file}
}
\value{
BStringSet object
}
\description{
read aligned sequences in phylip format
}
\author{
ygc
}

