% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readNumFilter}
\alias{readNumFilter}
\title{readNumFilter}
\usage{
readNumFilter(files, percentage = 2)
}
\arguments{
\item{files}{files}

\item{percentage}{percentage}
}
\description{
filter out read numbers below specific percentage
}
\author{
ygc
}

