% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/removeSeq.R
\name{removeSeq}
\alias{removeSeq}
\title{removeSeq}
\usage{
removeSeq(aln, idx)
}
\arguments{
\item{aln}{alignment}

\item{idx}{index}
}
\value{
alignment
}
\description{
remove sequence(s) in alignment by the specific index.
}
\author{
ygc
}

