% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NGS.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{NGS} instance}
}
\value{
print info
}
\description{
show method for \code{NGS} instance
}
\author{
Guangchuanbg Yu
}

