% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{toCodon}
\alias{toCodon}
\alias{toCodon,NGS,ANY-method}
\title{toCodon method}
\usage{
toCodon(x, ...)

toCodon(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameter}
}
\value{
matrix
}
\description{
toCodon method

toCodon method
}
\author{
Guangchuang Yu
}

