% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toNGS}
\alias{toNGS}
\title{toNGS}
\usage{
toNGS(fasta, start = 1)
}
\arguments{
\item{fasta}{fasta file}

\item{start}{start position, for translation}
}
\value{
NGS object
}
\description{
read fasta sequences and construct NGS object
}
\author{
Yu Guangchuang
}

