% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NGS.R
\docType{methods}
\name{toPosition}
\alias{toPosition}
\alias{toPosition,NGS,ANY-method}
\title{toPosition method}
\usage{
toPosition(x, ...)

toPosition(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameter}
}
\value{
list
}
\description{
toPosition method

toPosition method
}
\author{
Guangchuang Yu
}

