% Generated by roxygen2 (4.0.2): do not edit by hand
\name{translateSeq}
\alias{translateSeq}
\title{translateSeq}
\usage{
translateSeq(seq, start = 1, print = TRUE)
}
\arguments{
\item{seq}{DNA sequence}

\item{start}{translate start position}

\item{print}{TRUE OR FALSE}
}
\value{
character
}
\description{
translate DNA sequence to protein sequence
}
\author{
ygc
}

