% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treeAnno.phyDat}
\alias{treeAnno.phyDat}
\title{treeAnno.phyDat}
\usage{
treeAnno.phyDat(inTree, outTree = "out.nwk", anno, plot = FALSE)
}
\arguments{
\item{inTree}{input tree in nwk format}

\item{outTree}{output tree file}

\item{anno}{phyDat object}

\item{plot}{logical}
}
\value{
tree
}
\description{
tree annotation of sequence substitution
}
\author{
Yu Guangchuang
}

