% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeAnno.pml.R
\name{treeAnno.pml}
\alias{treeAnno.pml}
\title{treeAnno.pml}
\usage{
treeAnno.pml(pmlTree, outTree = "out.nwk", translate = TRUE,
  removeGap = TRUE, plot = FALSE)
}
\arguments{
\item{pmlTree}{tree in pml format, output of optim.pml (phangorn)}

\item{outTree}{output tree file}

\item{translate}{logical}

\item{removeGap}{logical}

\item{plot}{logical}
}
\value{
tree
}
\description{
tree annotation of sequence substitution by comparing to parent node
}
\author{
Yu Guangchuang
}

