% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treeAnno}
\alias{treeAnno}
\title{treeAnno}
\usage{
treeAnno(inTree, outTree = "out.nwk", anno, plot = FALSE)
}
\arguments{
\item{inTree}{input tree in nwk format}

\item{outTree}{output tree in nwk format}

\item{anno}{site annotation, e.g. N-glyco sites}

\item{plot}{boolean variable, plot or not.}
}
\value{
tree
}
\description{
annotating phylogenetic tree with functional site mutation
}
\author{
Guangchuang Yu
}

