% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAlign.R
\name{writeAlignedSeq}
\alias{writeAlignedSeq}
\title{writeAlignedSeq}
\usage{
writeAlignedSeq(aln, output)
}
\arguments{
\item{aln}{alignment}

\item{output}{out file}
}
\description{
write aligned sequence to fasta file
}
\author{
ygc
}

