% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeConsensus}
\alias{writeConsensus}
\title{writeConsensus}
\usage{
writeConsensus(aln, output)
}
\arguments{
\item{aln}{alignment}

\item{output}{output fasta file}
}
\description{
write consensus sequence to fasta file
}
\author{
ygc
}

