% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addKeywordFASTA}
\alias{addKeywordFASTA}
\title{addKeywordFASTA}
\usage{
addKeywordFASTA(fasta, keyword, suffix = TRUE)
}
\arguments{
\item{fasta}{fasta file}

\item{keyword}{keyword to be added}

\item{suffix}{logical, true add suffix, false add prefix}
}
\description{
add keyword in description line of FASTA file
}
\author{
ygc
}

