% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addKeywordFASTA.R
\name{addKeywordFASTA}
\alias{addKeywordFASTA}
\title{addKeywordFASTA}
\usage{
addKeywordFASTA(fasta, keyword, by, outfile = NULL)
}
\arguments{
\item{fasta}{fasta file}

\item{keyword}{keyword to be added}

\item{by}{one of suffix, prefix or mapping}

\item{outfile}{output file}
}
\description{
add keyword in description line of FASTA file
}
\author{
ygc
}

