% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoReport.R
\name{autoReport}
\alias{autoReport}
\title{autoReport}
\usage{
autoReport(contig.folder, ref.folder, name.file, out.folder = "output",
  filter = FALSE, percentage = 1, markAmbiguous = TRUE)
}
\arguments{
\item{contig.folder}{contig folder
files in contig.folder should have name contains _S\\d+[HNMP]+, or it will be move to unknown folder
should contains for example: XXX_S111NA_454.fas, XXX_S111NA_Contigs.fna and XXX_S111NA_mira_delX.fasta}

\item{ref.folder}{reference folder
should contains files like XXX__S111NA_Ref.fas}

\item{name.file}{name file}

\item{out.folder}{output folder}

\item{filter}{logical}

\item{percentage}{filter percentage}

\item{markAmbiguous}{makr ambiguous with * or not}
}
\description{
auto report sequence alignment and consensus sequence
}
\author{
ygc
}

