% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{autoTranslate}
\alias{autoTranslate}
\title{autoTranslate}
\usage{
autoTranslate(seq)
}
\arguments{
\item{seq}{DNA sequence}
}
\value{
list
}
\description{
auto determine the best translate start postion (offset 1,2 or 3),
and perform translation
}
\author{
ygc
}

