% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fas2phy}
\alias{fas2phy}
\title{fas2phy}
\usage{
fas2phy(fas, type = "DNA", outfile = "out.phy", collapse = FALSE)
}
\arguments{
\item{fas}{aligned sequences in fasta format}

\item{type}{one of DNA, RNA or AA}

\item{outfile}{output file}

\item{collapse}{collapse to 1 line or not}
}
\description{
convert fasta to phylip format
}
\author{
ygc
}

