% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getConsensus.R
\name{getConsensus}
\alias{getConsensus}
\title{getConsensus}
\usage{
getConsensus(seq.df)
}
\arguments{
\item{seq.df}{sequence data.frame, sequence is stored in row}
}
\value{
consensus sequence in character vector
}
\description{
determine the consensus sequence
}
\author{
ygc
}

