% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/hamming.R
\name{hamming}
\alias{hamming}
\alias{hamming.BStringSet}
\alias{hamming.character}
\alias{hamming.characterVector}
\title{hamming}
\usage{
hamming(seq1, seq2, indel = FALSE, ...)

\method{hamming}{BStringSet}(seqs, indel = FALSE, ...)

\method{hamming}{character}(seq1, seq2, indel = FALSE, ...)

\method{hamming}{characterVector}(seq1, seq2, indel, ...)
}
\arguments{
\item{seq1}{sequence 1}

\item{seq2}{sequence 2}

\item{indel}{include indel or not}

\item{...}{additional parameter}
}
\value{
numeric
}
\description{
hamming
}
\author{
ygc

ygc

ygc
}

