% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pmlToSeq}
\alias{pmlToSeq}
\title{pmlToSeq}
\usage{
pmlToSeq(pml, includeAncestor = TRUE, output = "seq.fa")
}
\arguments{
\item{pml}{pml object}

\item{includeAncestor}{logical}

\item{output}{out fasta file}
}
\value{
XStringSet
}
\description{
convert pml object to fasta file or XStringSet object
}
\author{
ygc
}

