% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printAlignedSeq}
\alias{printAlignedSeq}
\title{printAlignedSeq}
\usage{
printAlignedSeq(aln, window = 80, printAmbiguous = FALSE)
}
\arguments{
\item{aln}{alignment}

\item{window}{window}

\item{printAmbiguous}{logical}
}
\description{
print aligned sequences
}
\author{
ygc
}

