% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readNumFilter.R
\name{readNumFilter}
\alias{readNumFilter}
\title{readNumFilter}
\usage{
readNumFilter(files, percentage = 2)
}
\arguments{
\item{files}{files}

\item{percentage}{percentage}
}
\description{
filter out read numbers below specific percentage
}
\author{
ygc
}

