% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setNGS_}
\alias{setNGS_}
\alias{setNGS_,NGS,ANY-method}
\title{setNGS_ method}
\usage{
setNGS_(x, ...)

setNGS_(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameter}
}
\value{
NGS object
}
\description{
setNGS_ method

setNGS_ method
}
\author{
Guangchuang Yu
}

