% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(out.folder, name.file)
}
\arguments{
\item{out.folder}{out folder}

\item{name.file}{name file}
}
\description{
summarize the result
}
\author{
ygc
}
