% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NGS.R
\docType{methods}
\name{toAA}
\alias{toAA}
\alias{toAA,NGS,ANY-method}
\title{toAA method}
\usage{
toAA(x, ...)

toAA(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameters}
}
\value{
matrix
}
\description{
toAA method

toAA method
}
\author{
Guangchuang Yu
}

