% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/NGS.R
\docType{methods}
\name{toCodon}
\alias{toCodon}
\alias{toCodon,NGS,ANY-method}
\title{toCodon method}
\usage{
toCodon(x, ...)

toCodon(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameter}
}
\value{
matrix
}
\description{
toCodon method

toCodon method
}
\author{
Guangchuang Yu
}

