% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{toPosition}
\alias{toPosition}
\alias{toPosition,NGS,ANY-method}
\title{toPosition method}
\usage{
toPosition(x, ...)

toPosition(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameters}
}
\value{
list
}
\description{
toPosition method

toPosition method
}
\author{
Guangchuang Yu
}

