% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NGS.R
\docType{methods}
\name{toString}
\alias{toString}
\alias{toString,NGS,ANY-method}
\title{toString method}
\usage{
toString(x, ...)
}
\arguments{
\item{x}{NGS object}

\item{...}{additional parameters}
}
\value{
matrix
}
\description{
toString method
}
\author{
Guangchuang Yu
}

