% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeConsensus}
\alias{writeConsensus}
\title{writeConsensus}
\usage{
writeConsensus(aln, output)
}
\arguments{
\item{aln}{alignment}

\item{output}{output fasta file}
}
\description{
write consensus sequence to fasta file
}
\author{
ygc
}

